Exit with the right status when build fails
If the build fails, exit with the failure code (1). This is done by
throwing any errors from the build instead of just logging them to
stdout like before.

The correct exit code is important when running Gatsby in a CI build,
for example. You want the build to fail so you don't deploy the site,
if it didn't build successfully.