Remove react-hmre from Dev HTML Compiler Config (#305)
* Remove react-hmre from Dev HTML Compiler Config

Also fixes a bug where react-hmre would get added twice to the
DEFAULT_BABEL_CONFIG as there are two webpack compilers for the
development server. This was mutating the const object when it should be
creating a new default for each run. Now a new object is returned
instead.

Fixes #304

* Introduce a develop-html stage to webpack.config