edited some header links that didn't work (#2520)
* edited some header links that didn't work

? marks might have made the links non-functional. Also I fixed a typo

* The How to contribute section has some incomplete parts

I had to take out some FreeCodeCamp stuff and don't know what to replace it with...Also, are all these instructions redundant? It seems like we can just link to a github article on how to create a PR, and then link to the real How to Contribute page that already exists?? https://www.gatsbyjs.org/docs/how-to-contribute/

Same with running the Guide locally on computer. Isn't that info already in https://www.gatsbyjs.org/docs/how-to-contribute/?