Fix page rendering issue [breaking] (#7968)
Disabled componentDidUpdate for now, because when a user goes to any /community link, filtered or not, it errors and does not render anything due to a race condition with componentDidMount.

Was using componentDidUpdate primarily to keep up with back and forward navigation regarding the filters, will fix it sometime this week. Everything other than back forward filter updates should work correctly, even if a user is opening a /community link on a new tab.