Create using-gatsby-image.md (#8131)
* Create using-gatsby-image.md

Issue #8103: This is for the second task, "using gatsby-image to prevent image bloat". I did my best to make sure the links from the original article were still present.

Not sure how to navigate the fact that in the original article, it's specifically Kyle talking, so right now everything's still in first-person. I also cut the last line, which was leading to the next section of the article so wouldn't be relevant here (unless I should have included that next section as well?).

Let me know if there's anything you'd like me to change - this is my first-ever contribution to an open source project (currently in a web dev bootcamp), so I'm sorry if I screwed anything up!

* Update using-gatsby-image.md

Updated with suggested changes!

* just added one more link

* fixes