feat(examples): add gatsby + shopify example (#11725)
## Description

In an effort to reproduce caching issues, I created a simple Gatsby + Shopify example. Similar to gatsbyjs/store.gatsbyjs.org, this uses the same Gatsby Swag store account/instance in Shopify (which also commits the SHOPIFY_ACCESS_TOKEN to version control!).

This one is more read-only, and renders a list of products, and then product detail with a thumbnail switcher.