feat: add contributing docs re-organization (#29)
The Gatsby.js docs, while wonderful, are light on coding instructions for contributors. With dedicated sections in the Contributing docs, it will be easier to learn how to clone the repository, where to find certain things like docs or plugins, how to run tests, and insider infrastructure knowledge required to submit a successful pull request on the first try.

A lot of this information already lives on the "How to Contribute" page, but it's buried. It would help discoverability if this information lived on separate pages. Doing this means we can easily add a new section on "Setting up test environments", such as using Docker to hook up an easy Wordpress test installation.

I would love to hear your feedback as this is my first Gatsby RFC. I'm totally open to suggestions!