feat(gatsby-transformer-asciidoc): add pathPrefix handling for images, define supported file extensions and add "page-*" attributes from asciidoc file to AsciiDoc node (#13357)
## Description
This PR provides 3 additional features to the asciidoc transformer:

- Handling of pathPrefix for the imagesdir attribute
- Define which file extensions are supported by the transformer (e.g. .adoc and .ad)
- Add all asciidoc attributes starting with "page-" to the AsciiDoc node

### Handling of pathPrefix
The current gatsby-transformer-asciidoc plugin does not handle pathPrefix for the images used in the asciidoc file. It is common in asciidoc to set an imagesdir as an attribute like this:

      resolve: `gatsby-transformer-asciidoc`,
      options: {
        attributes: {
          imagesdir: `/images`,
        },
      }
The default value is /images

The PR checks if a pathPrefix is set and prefixes the defined imagesdir with it before Asciidoctor.js is called.

### Define supported AsciiDoc file types
In the current implementation only .adoc files are supported but there are multiple file extensions used for asciidoc like .ad or .asciidoc. The PR offers a way to define which fileExtensions are used in the project:

      resolve: `gatsby-transformer-asciidoc`,
      options: {
        attributes: {
          imagesdir: `/images`,
        },
        fileExtensions: [`ad`, `adoc`],
      }
The default values are ['adoc','asciidoc']

### Add all asciidoc attributes starting with "page-" to the AsciiDoc node
In other static site generators like middleman or awestruct, which I have used in the passed, it was common that the developer/blogger could define attributes inside the asciidoc file without the need of frontmatter. The common way was to define attributes that started with "page-" like:
`page-tags`, `page-category`, `page-permalink` etc.

This PR takes all of the defined page-* attributes and adds them to the AsciiDoc node and adds them under the pageAttributes value. The prefix "page-" will be removed from the key name. This way those page attributes can be queried via GraphQL and be used for creating the pages like this:

    {
        allAsciidoc {
           nodes {
              pageAttributes {
                  path
              }
           }
        }
    }