Quick fix to recipes discoverability (#23767)
* Quick fix

Should we list the official recipes out?

I'm not sure how to link to the community recipes individually. Where is that list? Or we could just share this in the umbrella issue and tweet it and ask people to PR in their own link.

* Update automated-recipes.md

* chore: format

* couple updates

thanks @laurieontech and @ascorbic!

* simplified PR

ok @laurieontech and @aishablake I think this will be the simplest quick fix to solve SEO issue

* Update relative link

Co-authored-by: gatsbybot <mathews.kyle+gatsbybot@gmail.com>
Co-authored-by: Aisha Blake <aisha@gatsbyjs.com>