feat(gatsby): Don't crash the build process when in preview mode (#33184)
* feat(gatsby): Don't crash the build process when in preview mode

This causes problems to a number of Gatsby Cloud customers. When doing preview builds, there's often nodes that are missing information as the user is still filling out the content form. We do want to crash production builds
in this case as we don't want 1/2 formed pages but for preview, we should error but not crash.

Fixes this Gatsby Cloud feature request https://gatsby.canny.io/gatsby-cloud/p/errors-shouldnt-necessarily-break-preview

The error page is very spartan but perhaps good enough for now.

* Add more to error & customize it for Preview

* make this conditional

* cloud sets booleans as strings...

* Better error html message

* Revert chunkSize change + make pretty structured errors

* Add the error name/stack to the generated error html page

* restore breaking prod builds

* just need stack

* Improve error

* fix lint error

* Fix typescript errors

* remove console.logs

* Add page data & component id to preview error page

* Update packages/gatsby/src/utils/worker/child/render-html.ts

Co-authored-by: Vladimir Razuvaev <vladimir.razuvaev@gmail.com>

* fixes for reviews

Co-authored-by: Vladimir Razuvaev <vladimir.razuvaev@gmail.com>